+++ tests/autotune/test_tune.py
@@ -91,7 +91,7 @@ def test_run_autotune_scvi_basic_mdata(save_checkpoints: bool, save_path: str):
 
 @pytest.mark.autotune
 @pytest.mark.parametrize("n_batches", [3])
-def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int):
+def test_run_autotune_scvi_no_anndata(n_batches: int, save_path: str):
     from ray import tune
     from ray.tune import ResultGrid
 
@@ -259,7 +259,7 @@ def test_run_autotune_scvi_with_scib_mdata(model_cls, metric: str, solver: str, 
 
 @pytest.mark.autotune
 @pytest.mark.parametrize("metric", ["iLISI"])
-def test_run_autotune_scvi_with_scib_ext_indices(save_path: str, metric: str):
+def test_run_autotune_scvi_with_scib_ext_indices(metric: str, save_path: str):
     from ray import tune
     from ray.tune import ResultGrid