+++ CHANGELOG.md
@@ -10,7 +10,7 @@ to [Semantic Versioning]. Full commit history is available in the
 
 #### Added
 
-- Add support for mudata during scib-metrics optimization, {pr}`3545`.
+- Add support for mudata during ray autotune, {pr}`3545`.
 
 #### Fixed